+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.2 - 2023-03-07
+### Added
+- Show Description option
+- Show Credits option
+
+### Changed
+- default lang based off plugin settings
+
 ## 2.0.1 - 2022-12-12
 ### Fixed
 - Styling issues